+++ Code/bardTagData.h
@@ -21,10 +21,6 @@ namespace bard {
 /**
  * \class TagData
  * \brief Value type to hold centre point, 4 additional (corner) points [0-3] and point ID.
- *
- * This could be generalised to "any number of additional points", but for now is hard coded
- * to exactly 4, as we are working with square tags. However, the ordering is unspecified
- * and in practical terms will depend on what your tag tracking library outputs.
  */
 class TagData {
 
@@ -33,11 +29,11 @@ public:
   TagData();
   TagData(const TagData& another);
   TagData(const int& pointId,
-          const cv::Point2d centrePoint,
-          const cv::Point2d cornerPoint0,
-          const cv::Point2d cornerPoint1,
-          const cv::Point2d cornerPoint2,
-          const cv::Point2d cornerPoint3
+          const cv::Point2f centrePoint,
+          const cv::Point2f cornerPoint0,
+          const cv::Point2f cornerPoint1,
+          const cv::Point2f cornerPoint2,
+          const cv::Point2f cornerPoint3
           );
   ~TagData();
   TagData& operator=(const TagData& another);
@@ -54,23 +50,23 @@ public:
   /**
    * \brief Return the centre of the tag, normally used as the tag 'location'.
    */
-  cv::Point2d GetCentrePoint() const;
+  cv::Point2f GetCentrePoint() const;
 
   /**
    * \brief Enables external clients to update this centre point.
    */
-  void SetCentrePoint(const cv::Point2d& point);
+  void SetCentrePoint(const cv::Point2f& point);
 
   /**
    * \brief Returns the stored corner point.
    * \param cornerPointId integer between 0 and 3 inclusive.
    */
-  cv::Point2d GetCornerPoint(const int cornerPointId) const;
+  cv::Point2f GetCornerPoint(const int cornerPointId) const;
 
   /**
    * \brief Sets the corner point.
    */
-  void SetCornerPoint(const int cornerPointId, const cv::Point2d& cornerPoint);
+  void SetCornerPoint(const int cornerPointId, const cv::Point2f& cornerPoint);
 
 private:
 
@@ -80,8 +76,8 @@ private:
   void CopyData(const bard::TagData& another);
 
   int          m_PointID;
-  cv::Point2d  m_CentrePoint;
-  cv::Point2d  m_Corners[4];
+  cv::Point2f  m_CentrePoint;
+  cv::Point2f  m_Corners[4];
 
 }; // end class
